% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_santa.R
\name{secret_santa}
\alias{secret_santa}
\title{Email Secret Santa List}
\usage{
secret_santa(participants, seed = 42)
}
\arguments{
\item{participants}{tibble containing secret santa participants (name, email, sock size)}

\item{seed}{random seed to generate the listing}
}
\description{
Email Secret Santa List
}
